+++ CHANGELOG.md
@@ -7,6 +7,10 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 
 ## [0.11.0] - TBD
 
+### Added
+
+- Added `run_with_context` method to `WriterWithContext` to enable custom writing operations to be performed
+
 ### Changed
 
 - The default method for `aligner` is now the 3-pass algorithm which has a more stable runtime and memory performance for large sequences. This may result in different optimal alignments